Fixed a bug in DoubleClickSupportedListView.
Added double-click support to FileListFragment. To do so, It takes advantage of DoubleClickSupportedListView.

In addition, a custom view is added to the fragment. This let us include the empty list text, however we have lost the ability to automatically show the ProgressBar when the Adapter is not set. To solve that we will include our so known ListFragmentCustomLayout class.